refactor: auto-split test_impact_model.py
